# backend/Models/openAPI.yml
openapi: 3.0.3
info:
  title: Brewtiful API
  version: 1.0.0
servers:
  - url: http://{{baseurl}}
...
paths:
  /api/Category/FullHierarchy:
    get:
      summary: Get all categories
      deprecated: false
      description: ''
      tags:
        - api/Category
      parameters:
        - name: Accept
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            '*/*':
              schema:
                type: object
                properties: {}
              examples:
                '1':
                  summary: Get all categories
                  value: "{\r\n    \"id\": {integer},\r\n    \"name\": \"{string}\"\r\n}\r\n{\r\n    \"id\": {integer},\r\n    \"name\": \"{string}\"\r\n}"
          headers: {}
      security: []
...
  /api/Category:
    post:
      summary: Invalid fields (400)
      deprecated: false
      description: ''
      tags:
        - api/Category
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
        - name: Accept
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required:
                - name
            example:
              name: Kava1
      responses:
        '400':
          description: ''
          content:
            '*/*':
              schema:
                type: object
                properties:
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                required:
                  - title
                  - status
                  - detail
              examples:
                '1':
                  summary: Invalid fields (400)
                  value: "{\r\n    \"title\": \"Bad Request\",\r\n    \"status\": 400,\r\n    \"detail\": \"A category with the name '{Name}' already exists.\"\r\n}"
          headers: {}
      security: []
...